Support variables with invalid names
gnm() converts automatically tables to data.frames, which calls
make.names() on the dimnames. This means the old names no longer
correspond to what can be used in the formula, but also that we
can use such variables without trouble. When both dimensions lack
a name, gnm chooses Var1 and Var2 for names (this does not happen
when only one dimension has no name, in which case it is NA).