Add support for static rcL and improve diagonal handling
Allow fitting a model with constant scores and phi using rcL().
The difference with fitting the same table with rc() is that
the table is not collapsed to a two-way table, which impacts the
indicators. It is better to add this feature to rcL() because
rc() can be kept simple, only working with two-way tables.

But internally assoc.rc() is called by rcL(), which requires
a few changes to handle three-way tables, and per-layer diagonal
parameters. Thus, all two-way table assoc() functions are updated
to use the same code for diagonal parameters, which now allows
using per-layer or unique parameters. Also fix a bug where reordering
of the diagonal parameters was not done, which means parameters were
*wrong*.