Fix calling gnm()
The solution in the last commit did not work correclty either. The new
solution is even simpler, and allows finer tuning of what objects
are substituted in the call. In particular, we include 'start' since
it will not be present when update() is called. 'tab' is assumed to be
present in the caller enviroment, which is wrong if it was converted
by us, but is good enough for now (real solution would imply creating
a custom update.rc() function).

This time, also change gnm() calls for base models, as they currently
don't hanle ..., which can be a problem e.g. if 'eliminated' is set,
since the number of parameters will not match.