Allow passing start=NULL to use random starting values
Occasionally, using completely random starting values can be useful,
in particular for complex models (e.g. with diagonal parameters)
that are very slow to converge, but may be very fast with good
random starting values. Use start=NA to mean "reasonable values",
and NULL to mean "random" (just like gnm). Ensure we always
convert start to NULL or a correct vector when it is passed as NA.