Stop using gnm in model$call for better update()
By setting model$call to use our functions instead of gnm(),
we do not need to substitute all the parameters for update()
to succeed. This is a much saner model, as anyway update.gnm()
cannot be expected to work very well (no base models, no call to
assoc()...).

This requires a few changes in the jackknife code, as update()
now preserves the se="jackknife" argument, which would lead to
infinite recursion if not changed.