Fix yrcskew model
Did not work because of changes between gnm 1.0-1 and 1.0-2: the
skew coefficient was automatically constrained to 0 (in gnmFit.R
and gnmTools.R). Use the original formula from the article, which
works and gives the same results as before (but still different from
the article).

Use the absolute value of phi to compute adjusted scores, since here
it is negative.