Identify dimensions and scores to match original ones in replicates
When bootstrapping or jackknifing, the dimensions are not guaranteed
to stay the same: the second axis can become the first, signs can reverse
if the first (refrrence) category moves slightly. This is particularly
a problem with bootstrap, where changes to the sample can be large. Thus,
detect what transformations can be applied to reduce the sum of squares
between original and replicated scores.

For the hmskew model, scores are not rotationnally identified: the convention
is that the first category has a null angle. This means it has no variance
on the y axis, but other points depend on its variations. Thus, rotate
replicated scores so that the sum of squares is the smaller.