Fix replicates when model fails
If gnm() fails badly, it may return NULL, or even raise an error. Detect
this to avoid blocking the whole process: we need use the original
model to start again. Also, return NA and rely on automatic recycling
rather than running assoc() just to find out the length of the value
to return.

Pass the tab argument when updating the original model: if missing,
the original value will be used and updated, which can be recursive if
the call contains something like tab[,,1:3].