Allow NAs in tables
This is useful to exclude specific cells, in particular the diagonal,
in a much more efficient way than including a full set of parameters.
Use na.rm=TRUE when computing margins. This has the downside that
marginal weights will not take into account the frequencies of NA
cells (this could be fixed internally by setting the cells to NA
afterwards...). When bootstrapping and jackknifing, we need to skip
cells with NA and handle them as if they were sampling 0s. Finally,
print() and summary() must handle NAs like gnm does.