Use procrustes orthogonal analysis for jackknife/bootstrap
The previous reflection/permutation of dimensions was not enough
since a rotation improves the stability of the scores. Procrustes
orthogonal analysis is also a simpler solution.

For hmskew and hmskewL models, we continue using the old solution
because dimensions must be handled in pairs (for now, only one is
supported, but...), and reflections are not allowed _inside pairs_.
Procrustes analysis does not bring any improvement except simplicity,
so for now it will do.

Add documentation about jackknife and bootstrap to ?se.assoc.