Split covmat into two parts to reduce memory use
With many layers, the full covariance matrix became very large,
and a model could require 600MB when saved to a .RData file. As
we never need the covariance accross layers, only put in the
main covmat phi for each layer, and normalized scores for the layers
they vary on. Add a adj.covmat array containing stacked covmats for
adjusted scores of each layer level. This dramatially reduces the
total size of the matrices, and makes the code simpler.

By the way, change the order of coefficients, so that they are grouped
by layer, and then per dimension, instead of starting with rows and
then columns. Change the order of name components too.