Remove one attempt when replicate fails
We already use the original base model and the default starting values,
no need to use random values which could converge to a wrong solution
and break the results.

Fix check for is.null(base) which was at the wrong if() block. Change
a warning() into a call to cat(), as it's really the intendend result,
and the anonymous function is not user-friendly (we show a real warning
at the end anyway).