Better starting values using custom SVD/EVD from base model
residSVD() from gnm is not enough for us since it does not work well
with NAs or perfectly fitted cells, and does not support layer effect
nor symmetric scores. It also uses a working residuals, which amount to
(observed - fitted)/fitted: this is good for correlation models, but
not for association models, where we want to decompose log(observed/fitted).
Use that by default since it gives much better results, and in particular
prevents failures to converge that happen in several cases with the default
residuals decomposition.

In hmskew and hmskewL, only specify starting values when nd.symm=NA: when
the symmetric association is RC, starting values are hard to find because
symmetric and skew-symmetric associations interact much and modify the scores.
Anyway, such models converge quickly even with random values.