Various fixes to plot.assoc
Fix handling of supplementary points that have no diagonal value: use
the mean value instead of 1, so that the chosen color is used instead
of full black. Allow passing pch and font by hand if wanted. Do not
set col="white" when pch does not have a border so that points are
not white. Use prettyNum() instead of sprintf() so that phi is
printed using getOption("outDec") instead of always ".".