Fix hmskew association order in some cases
The SVD returns the two dimensions in a random order since they have the
same singular values, and the current algorithm was not enough to fix all
the cases. The only reliable solution is apparently to reconstruct
the association and compare its sign with that of the original one.

By the way, clean up the code that rotates scores without affecting the
association: since the SVD always returns a null score for the first
category on one of the dimensions, we can align the first point on the
horizontal axis, positive side quite easily.