Allow NA rows/columns in residSVD/EVD()
A row or column full of NAs makes sense with RC-L models, for
example when one layer does not have this level, or when it has
too few cases so that model would not converge. In that case,
rowMeans() and colMeans() return NA, so we need to fill it with
an arbitrary value that will be replaced anyway as part of 'miss'.