Support tables with no dimnames and empty rows/columns
gnm automatically creates dimnames() and names(dimnames()) for us
if missing, but we only handled the second case. Now we also
support columns with only 0s or NAs, which are automatically skipped
by gnm (note that for symmetric models this still creates problems
if the resulting table is no longer symmetric).

Factor out the table checking logic to a new prepareTable() function
to reduce duplication.