Report AIC/BIC equal to 0 for saturated model
The version reported by AIC and extractAIC is not equal to 0
for the saturated model, which is not in line with the log-linear
modeling tradition. This also fixes the test in ?anoas, which was
not updated when summary.anoas() was changed not to use this version,
for consistency with summary.assocmod(), in commit 0d80518.