Add comment about scaling of supplementary scores
Only one constrain is needed to identify scores, so only center
them. OTOH, we need to divide them by sqrt(phi) to allow comparing
their sizes with standardized scores, and to plot them. Make
this clear in the code, and do not use [1,] so that if by mistake
a model with several layers ends up here, we get an error.