Fix identifying dimensions of hmskewL model (again)
Commit d099ccf was right in its explanation but the fix was incomplete. We
need to do as in hmskew() and compare the sign of the association described
by the normalized scores to that described by the original scores (lambda).
Move the per-layer operations to the loop above so that we do not need to
store lambda, and because it's cleaner that way.

Finally, the code changing the sign of the layer coefficients so that
the first layer is always positive was wrong: changing the sign of scores
does not affect the association they model; only one of the dimensions
must change sign (choose the second one since we standardize the position
of the first category on the first one). Add a test which failed before
this fix (the first test did not fail); correct results where checked
using a hmskew model.