Improve phi association coefficient in UNIDIFF
Compute the association coefficient by default, using marginal
weights. This does not affect layer coefficients, which are the
most used coefficients. Plot them the classic way, but add an
argument to use phi. Improve print() and summary() so that
the weighting is shown, and compute layer phi association coefficients
in summary() too.