Fix calling anova(..., test=)
When a test is requested, anova() did not work for unidiff
and assocmod objects, as some components were missing from the
result of summary(). Fix this by re-using values from
summary.gnm(), and add asGnm() to allow converting back to plain
gnm objects. Spotted by Heather Turner.

For now, explicitly passing 'dispersion' to summary() or anova()
is not supported. For unidiff, this would require changes since
the covariance matrix is extracted at fitting time: it would have
to be computed again with the new value of dispersion.